Change default of CfgParser::clear and use AtomUtils in desktop names reading.
Changed the default parameter of CfgParser::clear to realloc by default
as the only case when alloc will not be wanted is in the CfgParser
destructor. Clean-up and make use of AtomUtils in new desktop name
reading code, not sure if this works though.

+++ src/Config.hh
@@ -61,7 +61,7 @@ public:
         return (num >= _screen_workspace_names.size())?_screen_workspace_name_default:
                                                        _screen_workspace_names[num];
     }
-    void setDesktopNamesUTF8(char *names, uint length);
+    void setDesktopNamesUTF8(char *names, ulong length);
 
     inline int getScreenEdgeSize(EdgeType edge) const { return _screen_edge_sizes[edge]; }
     inline bool getScreenEdgeIndent(void) const { return _screen_edge_indent; }
